requirements.txt
rembg
tqdm
python-slugify
